fix: enhance logic
